+++ lib/src/block_parser.dart
@@ -11,6 +11,7 @@ import 'block_syntaxes/empty_block_syntax.dart';
 import 'block_syntaxes/header_syntax.dart';
 import 'block_syntaxes/horizontal_rule_syntax.dart';
 import 'block_syntaxes/html_block_syntax.dart';
+import 'block_syntaxes/link_reference_definition_syntax.dart';
 import 'block_syntaxes/ordered_list_syntax.dart';
 import 'block_syntaxes/paragraph_syntax.dart';
 import 'block_syntaxes/setext_header_syntax.dart';
@@ -35,6 +36,18 @@ class BlockParser {
   /// Index of the current line.
   int _pos = 0;
 
+  /// Starting line of the last unconsumed content.
+  int _start = 0;
+
+  /// The lines from [_start] to [_pos] (inclusive), it works as a buffer for
+  /// some blocks, for example:
+  /// When the [ParagraphSyntax] parsing process is interrupted by the
+  /// [SetextHeaderSyntax], so this structure is not a paragraph but a setext
+  /// heading, then the [ParagraphSyntax.parse] does not have to retreat the
+  /// reading position, it only needs to return `null`, the [SetextHeaderSyntax]
+  /// will pick up the lines in [linesToConsume].
+  List<Line> get linesToConsume => lines.getRange(_start, _pos + 1).toList();
+
   /// Whether the parser has encountered a blank line between two block-level
   /// elements.
   bool encounteredBlankLine = false;
@@ -50,6 +63,7 @@ class BlockParser {
     const HorizontalRuleSyntax(),
     const UnorderedListSyntax(),
     const OrderedListSyntax(),
+    const LinkReferenceDefinitionSyntax(),
     const ParagraphSyntax()
   ];
 
@@ -88,12 +102,19 @@ class BlockParser {
     return lines[_pos + linesAhead];
   }
 
+  /// Advances the reading position by one line.
   void advance() {
     _pos++;
   }
 
+  /// Retreats the reading position by one line.
   void retreat() {
     _pos--;
+  }
+
+  /// Retreats the reading position by [count] lines.
+  void retreatBy(int count) {
+    _pos -= count;
   }
 
   bool get isDone => _pos >= lines.length;
@@ -114,10 +135,26 @@ class BlockParser {
   BlockSyntax? get parentSyntax => _parentSyntax;
   BlockSyntax? _parentSyntax;
 
+  /// Whether the [SetextHeadingSyntax] is disabled temporarily.
+  bool get setextHeadingDisabled => _setextHeadingDisabled;
+  bool _setextHeadingDisabled = false;
+
+  /// The [BlockSyntax] which is running now.
+  /// The value is `null` until we found the first matched [BlockSyntax].
+  BlockSyntax? get currentSyntax => _currentSyntax;
+  BlockSyntax? _currentSyntax;
+
+  /// The [BlockSyntax] which is running before the [currentSyntax].
+  BlockSyntax? get previousSyntax => _previousSyntax;
+  BlockSyntax? _previousSyntax;
+
   List<Node> parseLines({
     BlockSyntax? parentSyntax,
+    bool disabledSetextHeading = false,
   }) {
     _parentSyntax = parentSyntax;
+    _setextHeadingDisabled = disabledSetextHeading;
+
     final blocks = <Node>[];
 
     // If the `_pos` does not change before and after `parse()`, never try to
@@ -128,22 +165,50 @@ class BlockParser {
     // 설명 - parse()를 실행하고도 _pos가 변하지 않으면 같은 BlockSyntax을 더이상 파싱하지 않는다.
     BlockSyntax? neverMatch;
 
+    var iterationsWithoutProgress = 0;
     while (!isDone) {
+      final positionBefore = _pos;
       for (final syntax in blockSyntaxes) {
         if (neverMatch == syntax) {
           continue;
         }
 
         if (syntax.canParse(this)) {
-          final positionBefore = _pos;
+          _previousSyntax = _currentSyntax;
+          _currentSyntax = syntax;
           final block = syntax.parse(this);
           if (block != null) {
             blocks.add(block);
           }
           neverMatch = _pos != positionBefore ? null : syntax;
 
+          if (block != null ||
+              syntax is EmptyBlockSyntax ||
+              syntax is LinkReferenceDefinitionSyntax) {
+            _start = _pos;
+          }
+
           break;
         }
+      }
+      // Count the number of iterations without progress.
+      // This ensures that we don't have an infinite loop. And if we have an
+      // infinite loop, it's easier to gracefully recover from an error, than
+      // it is to discover an kill an isolate that's stuck in an infinite loop.
+      // Technically, it should be perfectly safe to remove this check
+      // But as it's possible to inject custom BlockSyntax implementations and
+      // combine existing ones, it is hard to promise that no combination can't
+      // trigger an infinite loop
+      if (positionBefore == _pos) {
+        iterationsWithoutProgress++;
+        if (iterationsWithoutProgress > 2) {
+          // If this happens we throw an error to avoid having the parser
+          // running in an infinite loop. An error is easier to handle.
+          // If you see this error in production please file a bug!
+          throw AssertionError('BlockParser.parseLines is not advancing');
+        }
+      } else {
+        iterationsWithoutProgress = 0;
       }
     }